Continued progress with FullCalendar integration
It is important that we are able to cancel recurring events.

The data model will be such that recurring events look different from
standalone events.

I have set up some hard-coded state that will determine the PSQL data
model.  I think this is a good place to begin setting up the postgres
tables for events.